move window logic to its own file
